#377 - Module update on git pull : finish
- Rollback git pull on error
- Clean some unused code in MeveoModuleXXX layers

+++ meveo-admin/ejbs/src/main/java/org/meveo/service/custom/CustomEntityTemplateService.java
@@ -546,6 +546,10 @@ public class CustomEntityTemplateService extends BusinessService<CustomEntityTem
     }
     
     public void removeData(CustomEntityTemplate cet) {
+    	if (cet == null) {
+    		return;
+    	}
+    	
         if (cet.getSqlStorageConfiguration() != null && cet.getSqlStorageConfiguration().isStoreAsTable()) {
             customTableCreatorService.removeTable(SQLStorageConfiguration.getDbTablename(cet));
